Suggested clarification and extensions
Using the editing to include suggested additions and clarifications on specific sections of the specification, and to create discussion.
(as recommended by @ianbjacobs).

Initially these are just comments and questions to test the view of the group. I will add suitable text once the concepts or extra examples are agreed to.