Require Relying Party ID as input
Add a required input to `SecurePaymentConfirmationRequest` for the
relying party ID. If the requested credentials do not have a matching
RP ID, then the authentication will fail in the "Verifying an
Authentication Assertion" steps.

Before UI is shown, i.e. in "Steps to check if a payment can be made",
the input rpId is only checked to be a valid URL. This ensures that this
is a breaking change before any UI is shown. The actual validation of
the RP ID happens during authentication.

Issue #164