Add option to allow instrument icon download fail (#171)
Add option to allow instrument icon download fail

This adds an option to the SecurePaymentConfirmationRequest for the
developer to allow the credential to still be authenticated if the
payment instrument icon fails to download or decode. This is done by
adding an attribute `iconMustBeShown` to the
`PaymentCredentialInstrument`. A boolean is added to the response output
indicating whether the icon was shown.

The algorithms to check, display, and respond to a request are updated
accordingly.

Issue #125