[Spec] Remove outdated write-up from 11.3
This section still contained an idea outline (for salting credential IDs) from
when the spec was in its development phrase. The actual idea is in issue #77
and shouldn't be part of the spec itself.

Also added a short paragraph on a possible user mitigation for 11.3, creating
distinct user accounts.

See #143