[Spec] Add note on SPC opt-in, and misc clarifications
This PR adds a note about what SPC opt-in means today (required for any use of
SPC) versus in our preferred future (required for 3p use of SPC). It also
contains a set of small clarifications, e.g. updating the 'steps to silently
determine...' to acknowledge that an RP ID will likely be needed.