[Spec] Remove references to create in cross origin iframe (#281)
The WebAuthn specification now allows credential creation in a cross origin
iframe, as of https://github.com/w3c/webauthn/pull/1801. As such, SPC no longer
needs to override or add this behavior.

We do retain (for now) the ability for the `"payment"` permission policy to
allow credential create in an iframe, instead of requiring
`"publickey-credentials-create"`. This should be removed one day, but one
small step to unification at a time :).

There is a small web-compat issue here. In the case of no transient user
activation, the SPC specification used to throw a `SecurityError` error.
However WebAuthn throws a `NotAllowed` error instead. See Chrome bug
https://crbug.com/41484826

Fixes #267