[Spec] Add support for a 'details' string for PaymentCredentialInstrument
This change adds the ability for the SPC caller to specify a short
additional detail string for the PaymentCrednetialInstrument. This
allows for a richer UX, e.g. in the world of cards one could specify
a rich 'product' name for the displayName, and then add the last-four
digits and expiry in details:

displayName=Fancy Bank Platinum Card
details=****1234 01/26