[Spec] Introduce additional output states (#292)
This change expands the possible outcomes from the transaction confirmation
experience, to include the ability for the user to explicitly indicate that
they do *NOT* want to continue with the current payment. To achieve this
in a privacy-compliant way, this changes SPC behavior to *always* display
the transaction confirmation UX, whether or not any credentials matched.

The possible outcomes of SPC are now:

1. A signed cryptogram returned - user wishes to proceed with payment and has
   successfully verified their identity
2. NotAllowedError - user wishes to proceed with payment but either cannot OR
   does not want to use the passed-in credentials to do that
3. AbortError - user explicitly does not wish to continue with payment. The
   integrator should return to merchant checkout screen or handle otherwise
   sensibly.
4. OptOutError - as today, the user wishes for the RP to remove this credential.

Fixes #275