[Spec] Add support for displaying logos from facilitating payment entities (#294)
This change adds the ability for the SPC caller to specify a sequence of
'facilitating entity logos', which are intended to be logos from payment
entities (e.g., card network or issuer in the credit card space) that
can provide context for the authentication to the user. Logos are specified by
the caller in a descending order of importance.

In order to avoid constraining individual user agent interface design, the user
agent is free to show a subset (including the empty subset) of the input logos,
but it must respect the order of importance - that is, it can only show a logo
with index j in the list if it is also showing all logos with indices i < j.

This change is intended to be future compatible for a world where we may later
add additional information about the logos, including potential rendering hints
and accessibility information, but for now constrains itself to the minimal
possible support.

See #197