Rely on the IDL definitions for dictionary members. (#304)
Before this change dictionaries and their members were defined in IDL
blocks, then the members were redefined using `<dfn>` elements inside
`<dl>`. After this change the `<dfn>` and `<dl>`'s are removed; instead
the IDL definitions of the members are referenced.

Additionally, this change reuses the styles for the domintro class from
the WHATWG styles. Notably, this adds "For web developers
(non-normative)" as a heading to each domintro block.

See issue #302.